Moved DigitDecompExtend to decomp.py
