+++ kerngen/pisa_generators/basic.py
@@ -252,6 +252,7 @@ class KeyMul(HighOp):
     output: Polys
     input0: Polys
     input1: KeyPolys
+    input0_fixed_part: int
 
     def to_pisa(self) -> list[PIsaOp]:
         """Return the p-isa code to perform a key multiplication"""
@@ -268,7 +269,7 @@ class KeyMul(HighOp):
                 op(
                     self.context.label,
                     self.output(part, q, unit),
-                    input0_tmp(2, q, unit),
+                    input0_tmp(self.input0_fixed_part, q, unit),
                     self.input1(digit, part, q, unit),
                     q,
                 )
